backend: bug fix - Deletar apólice
Realizado correção do bug que estava ocasionando a deleção da seguradora realizar a deleção de uma apólice.
Além disso, foi adicionado uma verificação de um tratamento de exceção para não permitir a deleção de apólice caso ela já esteja associada a um Estágio ou Termo De Estágio.

+++ backend/estagio-poc/src/main/java/br/ufpr/estagio/modulo/service/ApoliceService.java
@@ -54,7 +54,20 @@ public class ApoliceService {
 	    Optional<Apolice> apoliceOptional = apoliceRepository.findById(apolice.getId());
 	    if (apoliceOptional.isPresent()) {
 	        Apolice apoliceExistente = apoliceOptional.get();
-	        apoliceRepository.delete(apoliceExistente);
+	        if (apoliceExistente.getEstagio() != null || apoliceExistente.getTermoDeEstagio() != null) {
+	        	throw new RuntimeException("Não é possível deletar uma apólice que esteja associada a um Estágio ou a um Termo de Estágio.");
+	        }else {
+	        	Seguradora seguradora = apoliceExistente.getSeguradora();
+	        	if(seguradora != null) {
+	        		List<Apolice> listaApolices = seguradora.getApolice();
+	        		if (listaApolices != null) {
+	        			listaApolices.remove(apoliceExistente);
+	        			seguradora.setApolice(listaApolices);
+	        		}
+	        	}
+	        	apoliceExistente.setSeguradora(null);
+	        	apoliceRepository.delete(apoliceExistente);
+	        }
 	    } else {
 	    	throw new RuntimeException("Não foi encontrado uma apólice com o ID informado.");
 	    }
